requirements.txt
python-telegram-bot==21.11.1
fastapi==0.115.12
uvicorn==0.32.1
aiosqlite==0.20.0
python-dotenv==1.0.1
itsdangerous==2.2.0
